+++ CHANGELOG.md
@@ -6,6 +6,11 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [Unreleased]
 
+## [1.2.0] - 2017-11-12
+### changed
+- Adds model stock.dispenser inheriting from model res.users.
+- Adds views for configure stock dispensers.
+- Adds button for activation/inactivation of dispensers.
 
 ## [1.1.0] - 2017-08-30
 ### added